Settings section: ensure that settings inside a revealer can be visible
Previously, the frame was only added to a section when there was a 'normal' row (without a revealer). This was most likely intended to prevent empty frames from looking bad, but it had the added side effect of not showing anything when there's only reveal rows - even if the revealer was shown. While a reveal section would generally be more appropriate (and would avoid this issue) it's still better to only hide the frame when we're sure there's nothing showing (ie all rows in the section are hidden).