Allow adding separators to the combo box widget
To use this feature, one need only pass the separator value as a kwarg and then any time that value is encountered in the list of options, that entry will be replaced with a separator. Note that the value must be in the first column for this to work. The value in the second column doesn't matter because it will be ignored.